+++ packages/tts-ggml/examples/cosyvoice-tts.js
@@ -25,6 +25,7 @@
 
 const fs = require('bare-fs')
 const path = require('bare-path')
+const proc = require('bare-process')
 const TTSGgml = require('../')
 const { createWav } = require('./wav-helper')
 const { setLogger, releaseLogger } = require('../addonLogging')
@@ -32,7 +33,7 @@ const { setLogger, releaseLogger } = require('../addonLogging')
 const COSYVOICE_SAMPLE_RATE = 24000
 
 const argv = global.Bare ? global.Bare.argv : process.argv
-const env = (global.Bare && global.Bare.env) || (typeof process !== 'undefined' && process.env) || {}
+const env = proc.env || {}
 const textArg = argv[2]
 const modelDirArg = argv[3]